Add metrics for dataflow nodes created during migration.
This commit adds a new counter DOMAIN_NODE_ADDED which
counts the number of nodes added in each domain by node type.
This allows us to verify changes to domains as a result of a
migration, such as when we replicate reader nodes.

This also adds one quality-of-life macro that allows tests to
retrieve metrics that match a subset of a metric/label
combination from a metrics dump.