Verify killing the controller reelects a leader in the primary region.
The existing `controller_in_primary` test is expanded to kill the
controller, verify that a new leader is elected and that the controller
is 1. not the original controller (as it is dead), 2. in the primary
region.

A modification is made to PersistentParameters to ignore the log
directory when verifying if the new config is compatible, as
different servers may have different log directories.