Make MirUnsupportedCondition include the expr
Add an `expression` field to MirUnsupportedCondition for the actual
expression that was unsupported, to aid in debugging instances of this
error that occur. I've elected to make the expression just a string
here, since it's really intended as a user- (or developer-) facing error
message, and carrying around an entire (serialized) AST feels pretty
heavy.