Implement a PostgreSQL WAL connector
This implements WAL logical replication to Noria.

The replication is using the native `pgoutput` plugin, that
is always present starting with PostgreSQL version 10.0.

The operation is similar to that of MySQL binlog, but is unable
to replicate schema changes due to the limitations of WAL, which
only streams row based events (INSERT,DELETE,UPDATE,TRUNCATE).

The replication also requires a user with `REPLICATION` privileges.

Fixes #[ENG-212]