logictest: Make column types optional
The original (sqlite) sqllogictest had column types as a required part
of every query I *think* because sqlite is dynamically typed, so there's
much less of a notion of a "data schema" and types are interpreted at
query-time. We kept column types around both for compatibility, and
because it's nice to be able to test that the correct data types get
returned. It'd be nice, however, to *not* specify the column types if
that's not something we're interested in testing (or, in my case, if
you're going to be parametrically generating queries and you don't want
to commit to always knowing what the types of the columns should be).
This commit makes column types optional, both (obviously) in the surface
syntax and in the Query data type itself, falling back on just turning
the Row returned by mysql into a Vec<DataType> if we don't have column
types at runtime.

Refs: ENG-240