Generate logictest scripts using query_generator
Extend the noria-logictest binary's `generate` subcommand to support
generating test scripts via a sequence of query operations passed to the
query generator, in addition to the existing functionality of generating
a test script via a seed script containing DDL and queries. This is
accomplished by encapsulating the overall generation options into a
`Seed` struct, which can be built from either a path to a seed script or
an instance of the query generator's Operations type.