Added logic to send non-reader domains exclusively to non-reader Workers.
During a migration, we now check if the domain is a reader domain by checking
to see if it has any reader node (reader nodes always belong on their own
domains).
If that's not the case, then we filter the Workers list and keep only the
non-reader Workers, and we pick one of those to send the domain to.