Fix handling of NULL values in comparisons
The semantics of NULL in SQL are tri-state; concretely, *any* comparison
with NULL, whether it's == or !=, always results in NULL, unless IS NULL
or IS NOT NULL is used. This means a `WHERE` condition that compares a
value to NULL should always fail. Previously, we'd just do `==`, which
used the Eq implementation for DataType which *does* consider
DataType::None equal to DataType::None (which is important for
reflexivity!) - now, we explicitly check for NULL first, unless the
operator is IS.