Factor out dataflow exprs into their own crate
Factor out the previous ProjectExpression type from within
`dataflow/ops/project` into its own crate, `dataflow_expression`, still
located within the dataflow/ directory, and rename it to
DataflowExpression. Putting this in its own crate is part of an ongoing
effort to have smaller self-contained crates in general (this particular
module has been getting pretty big as of late, and is going to be
getting even bigger in the near future), and renaming to
DataflowExpression is done in preparation to have the `filter` op use
this type in addition to just project, instead of its own more
constrained FilterVec type.