Eval the rest of the binary operators
Implement support in DataflowExpression for evaluating all
currently-unsupported binary operators, with the exception of LIKE and
its derivatives which will be implemented later. All of these operators
were previously (before ConditionExpression went away) only handled in
filters, which implemented its own version of boolean logic since it
only ever had to deal with boolean valued expressions - now, we have to
also deal with truthiness of other values, so that eg `1 and 2 == 1` and
`0 and 7 == 0` - this is implemented using a new `is_truthy` method on
DataType, which as far as I can tell (by issuing queries like `SELECT 1
WHERE <expr>` matches the truthiness semantics of the various MySQL
datatypes.