Add CaseWhen support to DataflowExpression
Add support to DataflowExpression for evaluating CASE WHEN expressions,
using the previously-added `is_truthy` method on DataType for
determining if the condition passes.

Unlike in the nom-sql AST, `then_expr`s are not optional here - when
lowering to DataflowExpression we just turn a missing then_expr to a
literal NULL. This follows the general principle that the lower the IR
gets, the fewer special semantics and corner cases it should have.