Get rid of FilterAggregate
Remove support for evaluating filters in aggregate (grouped) dataflow
nodes, and remove the FilterAggregate MIR node entirely, along with a
pretty significant amount of code whose only purpose was to support
filtered aggregates.

I don't have much insight into why support for filtered aggregates was
added as a separate node in the first place - as best I can tell, at the
time we couldn't aggregate on complex expressions (becasue support for
projecting those expressions in MIR before the aggregate hadn't been
added yet) and project didn't have the ability to project CASE WHEN
expressions yet (in fact, it was limited to either arithmetic or
literals). None of those restrictions are now present, and with the
unification of Expression and general simplification of a lot of the
very ad-hoc support for evaluating things the presence of all this
complexity is looking very dubious - not to mention that it was very
limited in the first place! (it could only support a column on one side,
and a literal on the other side). Add to that the number of lines this
diff manages to delete and this is all feeling like quite a good idea to
me.