Desugar IN to a series of ORs
When lowering nom_sql::Expression to DataflowExpression, desugar IN
expressions where the right-hand side is a list of expressions to a
series of = comparisons combined with OR, and NOT IN expressions to a
series of != comparisons combined with AND.

This is a relatively simple way of supporting IN in expressions (and
re-gaining it in filters, since they used to have it but lost it after
the refactor to make filters use DataflowExpression), but it may be the
case that especially for rather large lists the extra overhead of all
this nesting (and lots of pointer indirection!) becomes an issue.
Currently I'm of the opinion that if that happens, the way to fix it is
to switch DataflowExpression to a much simpler stack-machine-based
bytecode encoding, rather than attempt to give it native support for
IN (as always, low-level IRs should have as few semantics as possible!)
but we'll see.