Add a graphviz command to logictest
Add a new supported command to noria-logictest, which if currently
running against Noria will print the graphviz of the current query graph
to stdout.

Note that this:
- Doesn't run if the logictest has failed already (since we bail early
  on failures)
- Only prints out the *current* query graph

This means that if there's a query that returns the wrong results, the
way with this command to get its graph is to first make
a *statement* (with no expected results) to create the query in the
graph, then run graphviz.

This is implemented in a pretty messy way - since we support running
against mysql, we used to only pass around a mysql connection - now we
need to pass around an Option of the noria ControllerHandle as well.