Add an EXPLAIN [SIMPLIFIED] GRAPHVIZ statement
Add support for a new non-standard sql statement, EXPLAIN [SIMPLIFIED]
GRAPHVIZ, which will query noria for a (maybe simplified) graphviz
representation of the current query graph and report it back to the
client, in a resultset with a single column. The latter is done via a
new QueryResult::Meta enum variant, which leaves open the potential to
do more state-introspection stuff like this in the future.

Change-Id: I70f9d28f8351df7d8ffc1b46d5902d8e951dbb9f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/1112
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>