Make count(x) ignore null values of x
The SQL COUNT aggregate, when applied to an expression, is supposed to
omit rows where that expression evaluates to NULL. This commit
implements that behavior in the `apply_diff` for aggregates.

This unfortunately breaks the corresponding semantics for count(*),
since that's *not* supposed to ignore null rows, but since currently our
aggregates have to have an `over` column we have an early rewrite pass
that picks a random column for count(*) - meaning we won't count rows
where that column happens to be null! That'll be fixed next.