Make query-generator generate correct aggregates
When the query-generator generates queries that contains aggregates, we
also need to populate the GROUP BY of the query with all the columns
referenced by the query that *aren't* aggregated over - or else MySQL
will reject the generated query. This does that via a post-processing
pass after all query operations are populated.

In order to accomplish this I also had to move the `contains_aggregate`
function (and hence `is_aggregate`, which it calls) from noria_server's
query_utils module to nom_sql's analysis module - this *slightly* breaks
a separation of concerns principle I had been pushing towards where
nom-sql doesn't know anything about the *semantics* of the query, only
the syntax - but honestly I can't think of a better way of accomplishing
this. Maybe down the road we can break out *all* query analysis and
rewrite functionality into a separate crate that both query-generator
and noria-server can depend on.