Add more constrained operations to query_generator
When generating logictests to cover a particular bug or combination of
operations, it can be preferable to generate a smaller number of
queries, skipping over certain subsets of operations like individual
aggregates or separate filter operators that don't provide additional
coverage. Previously, however, certain specifications for query
generator operations, such as `aggregates` and *especially* `filters`,
would generate a *significant* - and sometimes even prohibitive, due to
blowing up the maximum number of columns in a mysql table - number
of permutations. This commit adds a few slightly smaller versions of
those two operations, constrained to only one aggregate in the case of
aggregates, and only one filter comparison op in the case of filters.