Allow generating BETWEEN filters
Add support to query_generator for generating BETWEEN (and NOT BETWEEN)
filters on two literal expressions. This is done by breaking out the
part of Filter that previously had the operator and the right-hand side
into a new FilterOp enum (but leaving behind extend_where_with), and
adding a new Between member to that enum with a single boolean negated
value. This is included in the ALL_FILTERS list, and also can be
specifically generated by passing `between_filters` on the command line
wherever Operations is supported.