Ignoring test for replication to non existent Worker.
This test is ignored because the Controller is not noticing that the Worker it
is trying to replicate domains to is no longer available.
Even with a low heartbeat/healthcheck interval, the test fails in our CICD
pipeline, but not locally.
We need a better fix, one that ideally does not depend on Worker timeout to mark
 a Worker as not healthy (maybe a way to tell the Controller to mark the Worker
 as such, but that's something we want only in the context of tests probably).