Add topk support to query_generator
Add a new TopK operation to query generator, which generates queries
with an ORDER BY (both descending and ascending) and a limit, and add a
simple generated logictest with topk.

This also required fixing a bug with the Display impl for OrderClause,
which didn't print the table names of columns that were being ordered
by, only the column name!