Properly handle extra zeros in Literal parse+print
The min precision for a Float literal in nom-sql is 30 - this means that
if we don't specify the precision ahead of time, we'd end up formatting
eg 1.5 as 1.5 with 29 trailing zeros - this would then fail to parse on
a round-trip due to overflowing the maximum size of an i32, which we
were using to parse the fractional part of a float literal. It's a
little silly that we were implementing float literal parsing ourselves
in the first place, since the rust stdlib has a (much more robust!)
parser in the form of the FromStr implementation of f64 - this commit
updates our float parsing to just use that instead, and also strips
trailing zeros when formatting float literals.