Allow generating IS NULL / IS NOT NULL filters
Add support to the query generator for generating IS NULL and IS NOT
NULL filters, implemented as a new variant of the FilterOp enum. This is
special (versus a regular old FilterRhs with an operator) because we
also need to annotate that values in the filter column need to be null
at least some of the time, which we do here using the `expect_value`
method on `TableSpec`.