Support postgres string literals and quoted identifiers in nom-sql.
This commit adds a feature flag for `postgres`, that, when enabled,
changes how quoted identifiers and string literals are parsed. This
allows building the nom-sql library for postgres.

The tests in this commit for postgres are not comprehensive and
are not apart comprhensive. Though, the format acts as a
sample of the intended format for the rest of the failing tests.