Add support for "expected failures" to logictest
When running logictest in verify mode over a set of input files,
classify those files into "expected passes" and "expected failures" -
where the latter case is test scripts that have a `.fail.test`
extension. Then, when *running* the scripts, *fail* the overall run if
any of the expected failures *pass*. This allows us to commit "broken"
logictests to the codebase, but get notified by a failing CI if we ever
inadvertently fix a broken logictest.