Implement PostgreSQL snapshot
This implements taking a snapshot of an existing PostgreSQL database
before WAL streaming replication takes place.

First the replication attempts to create a publication `FOR ALL TABLES`
on the primary. This operation requires super user priviliges and may
fail if those are not given to the user, in which case it is expected
that a publication named "noria" will be created out of band with the
proper tables added to it.

Second the replication will create a replication slot on the primary
with an exported snapshot. This step requires `REPLICATION` priviliges.

Last, using the exported snapshot, the recipe and the existing tables
are replicated into Noria. This requires `SELECT` priviliges on the
specified tables.

After the snapshot replication is done, WAL streaming replication takes
over from the point the snapshot was created.

Fixes #[ENG-213]