Allow generating logictests with delete statements
Add a new option, `--include-deletes`, to the logictest `generate`
subcommand, which when specified emits a sequence of DELETE
statements (with a configurable number of rows) after the first set of
queries, followed by the same queries again. This will allow us to test
the behavior of various operators as they receive negative deltas.

Since noria currently only supports DELETE statements with primary
keys (and the DeleteRow table operation has no way to be exposed via raw
SQL statements), this also requires adding support to query_generator's
table schema machinery for tracking and adding primary keys to tables.
That involves both *creating* primary keys for tables (which are just
unique columns which we remember are the primary key for the table) in
addition to *detecting* primary keys for CreateTableStatements that we
parse and pass to a GeneratorSpec.