Change to INNER due to aggregate limitations.
Changing this test to use an INNER pre-emptively due to known
limitations around left joins and the use of COUNT. Specifically, it is
known that if we write a SELECT where we select on COUNT *and* a column
from the right table with a LEFT JOIN and a GROUP BY (regardless of
what's in the group by), we break. We are planning to change DISTINCT to
use COUNT flow node under the hood, so we can handle row deletions
correctly (currently there is a bug that a row deletion removes the
value from the DISTINCT, even if the set of values still exist in
another row).

The bug in question has been deemed to be out of scope, and so this test
is being modified for the time being.