Change DISTINCT to use count under the hood
This commit changes distinct in a number of important ways. We still
create a Distinct MirNodeInner type, but when translating to flow, we
use Count. This commit also removes the distinct flow node entirely, as
it now serves no use.

This provides a few benefits:
1. We properly can support row deletion. Previously if a row was
deleted, it would cause a removal of those values from any DISTINCT
node. Now we simple decrement the count for that set of values, and only
omit it from the final output if the COUNT reaches 0.
2. We get partial materialization for DISTINCT by piggy backing on
COUNT, which already supports partial.

We add an extra column that holds the count itself when creating the
DISTINCT MirNode. This is done to be explicit. Count will create an
extra column to store the count in. This sets up a column to hold those
counts, used explicitely for implementing DISTINCT. This column gets
ignored in the final output because of it's absence from the projected
columns list.