Relax column_source constraints for LEFT JOINs
After thinking it over, it turns out `column_source` for joins doesn't
need to really care whether the join is a left join or not. The reasons
are slightly subtle:

It's not possible to do an upquery with the semantics of `IS NULL` [*],
which is the only place where we'd need to be careful as a left join; if
you theoretically could do this, the query

	SELECT * FROM a LEFT JOIN b WHERE a.col = ? AND b.col IS NULL

would be handled improperly by the `handle_upquery` logic (since it
would try and do another upquery for col IS NULL against b, instead of
doing the a \ b set difference operation).

The case where you index only columns on the RHS of a left join was
already handled fine prior to `column_source`, so it's okay to keep that
behaviour as is (forward the upquery to the RHS node and on_input will
deal with it).

And, the straddled case is equivalent to an INNER JOIN, since you can't
upquery for NULLs, and being in the straddled case implies you've
specified constraints on both LHS and RHS tables; you must, therefore,
have actual rows from both sides.

[*]: This isn't possible because you can only write `IS NULL`, and not
     `IS ?` (and then put `NULL` as a parameter); the latter isn't valid
     SQL. The former will just get compiled into a filter node by MIR.