Make packer builds not depend on release binaries
The `buildkite-agent artifact download` script will download the latest
version of an artifact, even one built by another build - this means
that if we're able to run the packer image build without first waiting
for the release binaries to build, it'll just download the latest built
version of the binaries, which will *significantly* speed up testing
packer image builds.

This *does* mean that the binary will have a different version than the
version annotation for the packer image itself - I'll fix that in a
follow-up commit.