Allow generating subqueries with query_generator
Add support to query_generator for generating queries with subqueries
either directly in joins or in CTEs (and then joined to), up to a given
depth of recursively nested subqueries.

Implementing this involved a fairly significant refactor of the way
options for generating queries are passed around in the query_generator,
since we need the ability to specify the QueryOperations for the
subqueries themselves. This is accomplished by wrapping up all the
options for generating a single query into a new QuerySeed struct, which
is the thing that's passed to `GeneratorState::generate_query` now,
instead of a vector of operations - it has that as a field, in addition
to a vector of a new Subquery struct, which specifies the position of
the subquery (join or CTE now, but IN can be added later) in addition to
a recursive QuerySeed for the subquery itself. The *user-supplied*
options for generating subqueries are now encapsulated into a new
`GenerateOpts` struct, which has an `into_query_seeds` method that
handles permuting the set of user-supplied operations and generating
recursively nested QuerySeeds up to the user-supplied subquery nesting
depth. That struct also derives Clap, which allows it to be specified on
the command-line as part of other subcommands (`noria-logictest
generate`, `query-generator benchmark`, etc.) using `#[clap(flatten)]`.
To support specifying the *number* of operations we want to permute
to (which we want to be a range of numbers) on the command line, this
finally adds a new `parse_bound_pair` function and an extension trait to
`launchpad` to turn that bound pair into an iterator over the range of
numbers.