Limit the number of unique columns generated
Previously, the only way to get a reference to a column with a specific
type when applying a QueryOperation was to use the
`fresh_column_with_type` method - this meant that every time we applied
eg a filter or a join, we'd add a new unique column to the relevant
table. With large enough permutations of query operations (which is
especially easy to get now that we have recursively nested subqueries)
this could result in very quickly hitting mysql's limit on the number of
columns in a table.

This commit adds a new method to TableSpec called
`some_column_with_type`, which tries to find an *existing* column with
the given type, only creating a new one if none already exists, and
updates a number of query operations that previously called
`fresh_column_with_type` to call that instead.