Add table to query in filter and aggregate
Similarly to what we do in joins, if we've added a column to filter on
or aggregate over, check to see if the query has no FROM tables and add
that column's tables if so. This makes both filters and aggregates work
when they're the only or first operation in a query.