Framework and initial data generation for the fastly demo.
This pulls parts of `noria-logictest` and `query_generator` to
generate random data to seed a test database, and writes the
data to a connected mysql database.

This is accomplished by parseing the CREATE TABLE ddl statements
used to setup the test database and user passed input parameters to
generate a `DatabaseGenerationSpec`, a struct defining what and how
we are going to generate test data. The `generate::load` function
handles turning that spec into insert statements and loading them
into the test database.