Revert "If NOT NULL is not specified for column, add NULL"
This reverts commit 39c761c7840e5403c66096a79a8a036423e9f17e.

The semantics of omitting nullability entirely are not quite the same as
those of having an explicit NULL constraint - specifically, if a column
is a primary key and missing a NULL then it will be given an implicit
NOT NULL constraint, but if it *has* a NULL mysql will throw an error.
This was breaking generated schemas for query-generator generated
logictests, so I'm reverting the change for now and am going to replace
it with a change that actually persists the NULL in the AST, and formats
that when displaying.