Track NULL on columns in the AST
Previously, we would throw away the NULL flag (constraint?) on columns
during parsing entirely, meaning that the parser didn't successfully
round-trip, causing issues when passing explicitly nullable columns
through things like the adapter. An initial pass at fixing this,
cf3f310a, just *always* wrote NULL to columns that were missing a
NotNull constraint - this, on the other hand, broke generated schemas
for eg noria-logictest that depended on the behavior of columns
implicitly getting a NOT NULL flag when eg used as a primary key. That
commit has been reverted, and this instead implements the correct
behavior of just not throwing away a Null column constraint when
parsing, which we can then format on the way back out. This was the only
reason the various column constraint parsers were returning Options,
too, so those can change to just return ColumnConstraint as well.