Compare logictest results type-insensitively
If we haven't specified types in the logictest (eitheer because it's a
manually written test and we don't care, or it's a hand-written test so
inferring the types is prohibitively complex), then we shouldn't
be *checking* to make sure the types are the same as what we've parsed.
This is especially important for generated tests, as mysql frequently
returns results as Text values even though the native type of the column
is actually something different - causing tests we've just generated to
immediately fail when re-run against mysql! This changes the result
checking in the logictest script runner to only do *direct* equality if
we actually have the expected column types in the script, and instead
convert the values to be the same type (by hopping through
`mysql::Value` and then `from_mysql_value_with_type`) before comparing
them if we don't.