Add to query tables in topk
Extend the work done for filters and aggregates in 9e155bec to topk - if
we add a new table to topk against and there are no tables in the query
yet, make sure that table gets added to the FROM clause of the query so
that it's a valid query. This makes generating queries with only the
topk operation work properly.