Fix bounds error in logictest parser error reporting
If we hit a parse error in logictest, we try to print the next 16
characters after the location of the parse error to try to give context
for the error - but if that parse error was hit less than 16 characters
from the end of the input file, that would panic with an out of bounds
error. This fixes that to clamp the 16 character limit to the length of
the string instead.