Add 981 new automatically-generated logictests
Add generated logictests for permutations of all "small" query generator
operations (eg between_filters, but not filters, or avg but not
aggregates) up to 3 operators combined per test, all to a new
`logictests/generated` directory.

Almost all of these tests currently fail (in fact, only 75 pass) due to
a number of bugs - for each of those, the test is marked as an expected
failure (with the .fail.test extension) and the JIRA ticket for the bug
documented by that failing test is documented in a comment at the top of
the file. This grouping of multiple test failures into buckets was done
manually, so it's possible some tests have been mistakenly grouped
together, or that two distinct failures are actually due to the same
root cause - but generally speaking I feel fairly confident about the
bucketing here.

Not *all* operations were able to be effectively converted to generated
tests - specifically, the project_builtin operation generates *very
large* tests due to having a lot of sub-operations that it represents,
and if those tests don't time out outright they'll frequently hit the
upper limit on the number of columns in a single table in mysql - that
issue is left to fix at some point in the future.

Currently these tests take at least an hour to run in CI (they may take
quite a bit longer, but they timed out after an hour) so this commit
does *not* enable them there (noria-logictest does not crawl directories
recursively) - as a next step, I'm going to be adding this to a nightly
build.