Don't run js-client tests unless js-client is changed
Don't run the tests for the js-client (including building the
noria-server docker image, which is *very* slow) unless the js-client
itself changes. This could result in us accidentally breaking the
js-client, so may want to be revisited if that ever becomes higher
priority, but for now the js-client is relatively low priority so this
isn't as important.