Fix all clippy lints in noria-server
Fix all clippy lints in the noria-server crate, via a combination of
automatic fixes (via `--fix -Z unstable-options`) and manual ones.

This crate is quite large (I would say too large!) so there's quite a
bit going on here - though almost all of it is quite mechanical and (I
would say) rather uncontroversial - lots of things like unnnecessary
clones, places where we can use `matches!`, etc.

The main place here where a decision had to actually be made is wrapping
the `request` field of the `WorkerRequestKind::DomainRequest` enum
variant in a `Box`, due to the `clippy::large_enum_variant` lint - that
particular field is quite large, so it makes sense to store it behind a
heap reference rather than storing it as part of the enum (since all
variants pay the cost for it). It's likely this actually *helps*
performance (due to more things being able to fit in a cache line), but
it's worth scrutinizing this a little further (perhaps a bit down the
road) to make sure this doesn't hurt us in any way.