Allow type_complexity on an iterator in noria-psql
Allow the clippy::type_complexity trait on the impl of IntoIterator for
noria_psql::Resultset. I'm not really sure that factoring out this type
into a top-level type alias would get us much, since it's tightly
coupled to the impl.

This is currently the only failing clippy lint in this crate, so I've
also enabled clippy in CI for it