Clippy the entire workspace in CI
Now that clippy is passing on all crates in the repo, we can replace the
piecemeal running of clippy on specific subcrates in the "check clippy"
step in CI with a single step that checks clippy on the entire
workspace.

Note that by default this only includes lib and binary targets, not test
targets - at some point in the future it might be nice to also run
clippy on tests, but that's omitted here for now.