Get rid of separate `cargo check` steps
`cargo clippy` is a strict superset of `cargo check`, so it's pointless
- and very slow! - to run it on separate sub-crates when we're already
running `cargo clippy` on the entire repository. So let's just get rid
of all the places where we're running a separate cargo check.

Also, when this changed from `build` to `check` the relevant entries in
the top-level buildkite.yaml didn't change, so this gets rid of those as
well.