Run all tests in a single build step
Running tests for separate sub-crates in the workspace was a
reasonable *idea*, and it is *kinda* nice to get separate CI statuses in
github for different crates failing their tests, but in practice the
amount of actual work that is *shareable* between the individual test
steps is way less than would be expected even with sccache, meaning that
the combination of all these steps takes a *huge* amount of time to run.
For the sake of cutting down on our CI time, let's just run all the
tests (with the exclusion of the clustertest crate, which doesn't
currently run in CI, and with the integration_serial module running
separately limited to one thread) for all the crates in the workspace in
one big CI step, combining together the docker-compose config for things
like noria-client that need zookeeper and taster, which needs redis,
into a single big docker-compose file for tests.