Run cargo-audit in CI
Add a pair of steps to CI to build a docker image containing cargo-audit
and run it against our dependencies whenever Cargo.lock changes, to
alert us if we ever depend on an insecure package either transitively or
intransitively.

Running this only on Cargo.lock changes *does* admittedly miss *new*
security advisories in existing dependencies, but this is a conscious
change so that failed builds on branches can only be triggered by a user
change (adding or changing a dependency) - my plan is to also have the
nightly plugin indiscriminately run `cargo audit` every time it runs.