Refactor TableSpec to include how to generate each column's data.
We introduce a new enum `ColumnGenerationSpec` and wrap the
`SqlType` and each column's `ColumnGenerationSpec` into a new
struct of column infromation: `ColumnSpec`. This ColumnSpec can
be used to generate data for a column. The existing `generate_data`
function behavior remains the same.

This commit includes several borrow-checker related fixes as
`TableSpec` is now mutable when generating data as certain
generators maintain internal state for how to generate the
next value.

A future change will support changing a column in a `TableSpec`
to a new `ColumnGenerationSpec`. Then, when `generate_data` is
called, this new spec will be used to `generate_data`.