Instrumented Prometheus for all Noria adapters.
Right now, this commits adds a Prometheus recorder and exporter with a default
configuration, which already sets up a server (at 0.0.0.0:9000) exposing
Prometheus metrics.
This should be good enough as a first step towards instrumenting the Noria
adapters. In the future, we can:
1. Move the metric recorders (now present inside the noria/server crate) to its
own crate, to be used by any other standalone service we want to implement in
the future.
2. Make the metrics instrumentation more customizable within the Noria adapters
(i.e., setting a different IP:PORT for the metrics to be exposed).