Refactor process for base table
Refactor the `process` operation for base tables, and fix a logic bug.

The logic bug was discovered when testing binlog replication and is
as follows:

Assuming a table has a row with the key 1 and an update arrives to this
row via binlog for a non-key column. Replicator translates it to `DeleteRow`
followed by `InsertRow`, and in the existing `process` loop `DeleteRow` would
clear the value of `current` but the `Insert` operation would fail because it
is being compared to the `was` value which is still there. This will result in
a negative record being emmited properly, but not the positive record.

This is also true for `DeleteByKey` operation.

The current change makes the whold flow much more readable, and similiar bugs
more difficult to introduce.

Fixes #[ENG-351]