Fix a logic bug in process that was not properly tracking key column updates
The current implementation of record processing only checks the currently
persisted row value for a given key, hence if an update were to change the key
column, subsequent operations would not be properly applied. This change keeps
track of local changes, not only persisted changes, in order to properly compute
the records.