Add a "recursive" flag to noria-logictest
Add a new "recursive" flag to all noria-logictest subcommands that take
a set of files as input (verify and parse, right now) via a
`#[clap(flatten)]`ed InputFileOptions struct, which uses the `walkdir`
crate to run all test files in a directory *recursively* if set. This
will be used to run the big logictest directory from sqlite in the
nightly builds.