Add readyset-framework-testing project
This adds what is essentially a clone of the vitess-framework-testing
repo, modified to be scrubbed of vitess/ps specific language, and
switched over to be ReadySet specific. The README has been
updated to give direction on how to test against a locally run ReadySet.

This also removes numerous helpers that were used to generate sequence
tables, and update vschema for a vitess specific testing scenario. We
don't need that functionality at all, so I've removed it entirely.